move create_users_parser.rb inside of customers
