+++ app/jobs/update_db_job.rb
@@ -3,7 +3,7 @@ class UpdateDbJob < ApplicationJob
 
   def perform(*args)
     logger.info { "I'm starting jobs" } 
-    CreateUsersParser.update_customer_db
+    Customer.update_customer_db
     Alert.check_customers_with_negative_acount
     logger.info { "OK I'm done now" }
   end
