+++ spec/jobs/update_database_job_spec.rb
@@ -3,7 +3,7 @@ describe UpdateDbJob do
   # to do!
     
   it "calls the update_customer_db method once the job is called" do
-    # expect(CreateUsersParser).to receive(:update_customer_db).at_least(1).times.and_call_original
+    # expect(Customer).to receive(:update_customer_db).at_least(1).times.and_call_original
     # UpdateDbJob.perform_later 
   end
